Renamed app barrel file
