+++ lib/main_staging.dart
@@ -1,4 +1,4 @@
-import 'package:zombie_conga_flame/app/app.dart';
+import 'package:zombie_conga_flame/app/app_barrel_file.dart';
 import 'package:zombie_conga_flame/bootstrap.dart';
 
 void main() {
