+++ test/app/view/app_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:zombie_conga_flame/app/app.dart';
+import 'package:zombie_conga_flame/app/app_barrel_file.dart';
 
 void main() {
   group('App', () {
